% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_relative_abundance.R
\name{calc_relative_abundance}
\alias{calc_relative_abundance}
\title{Relative abundances in phyloseq object}
\usage{
calc_relative_abundance(ps)
}
\arguments{
\item{ps}{a whole phyloseq object of taxonomic produced by the phyloseq::phyloseq() function.
ps is named thus as this is common shorthand for a phyloseq object in the field.}
}
\value{
an R dataframe of microbiome composition data containing sample ID and all other
provided metadata, taxonomic information, abundance values, and relative abundance values.
Abundance values of NA are dropped.
}
\description{
Starting from a phyloseq object, this applies the psmelt function to convert data into an R dataframe and subsequently calculate relative abundance of each taxonomic unit within its sample.
}
\examples{
data(esophagus, package = "phyloseq")
data(GlobalPatterns, package = "phyloseq")
data(soilrep, package = "phyloseq")
calc_relative_abundance(esophagus)
calc_relative_abundance(GlobalPatterns)
calc_relative_abundance(soilrep)
}
