% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esophagus.R
\docType{data}
\name{esophagus}
\alias{esophagus}
\title{esophagus Dataset}
\format{
A phyloseq object
}
\source{
phyloseq package
}
\usage{
esophagus
}
\description{
Small 16S rRNA data from just 3 human subjects' esophaguses, 1 each from 3 subjects, as part of the phyloseq package.
}
\keyword{datasets}
