% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalPatterns.R
\docType{data}
\name{GlobalPatterns}
\alias{GlobalPatterns}
\title{GlobalPatterns Dataset}
\format{
A phyloseq object
}
\source{
phyloseq package
}
\usage{
GlobalPatterns
}
\description{
16S rRNA data from 25 environmental seamples, as part of the phyloseq package.
}
\keyword{datasets}
