% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilrep.R
\docType{data}
\name{soilrep}
\alias{soilrep}
\title{soilrep Dataset}
\format{
A phyloseq object
}
\source{
phyloseq package
}
\usage{
soilrep
}
\description{
Small 16S rRNA data from 24 soil microbial communities under four treatments, as part of the phyloseq package.
}
\keyword{datasets}
